CHanged entry point. Updated styles.
