+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "just-for-today",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "app/App",
   "scripts": {
     "start": "npx expo start --tunnel",
     "android": "expo start --android",
